Make nova-placement-api default config less strict
Patch nova to enable execution of nova-placement-api if nova.conf
doesn't exist in the typical location of /etc/nova/nova.conf
and OS_PLACEMENT_CONFIG_DIR is not set. In our case with the
OpenStack snap for nova, we have two possible paths we may
pick nova.conf up from, based on what --config-file specifies.
The --config-file and --config-dir CLI options get parsed in
oslo.config (in the __call__() method of the ConfigOpts class).

Change-Id: I02e2a5d9df91eb8cbf36beb474809bc1e80fe80b
Closes-Bug: 1696830